Upgrade Plexus IO to 3.0.0 and clean-up legacy code
It may seem that there are lot of changes but actually most of them are:

* Rename `Java7FileAttributes` to `FileAttributes`. In the new version
  of  Plexus IO `FileAttributes` is replaced by `Java7FileAttributes`
* Remove calls to `Java7Reflector.isAtLeastJava7()`.
  `Java7Reflector` is deleted from the new version of Plexus IO
   because now at least Java 7 is required

The most significant change is made inside `ArchiveEntryUtils#chmod` - now
the Java version is at least 7 so together with the removal of the
call to `Java7Reflector.isAtLeastJava7()` the legacy code is removed as well.

Remove the `commons-io` dependency. Plexus IO 3.0.0 was updated to
use `commons-io` 2.5 so there is no longer a need to override
the transitive dependency.