Deprecate useJvmChmod
Prior to Java 7,  the `chmod` command was used to change the files mode.
`useJvmChmod` was a way to force the usage of the JVM instead of
the external command.

Now Java 7 is the minimum required version and no external commands
are needed to change the files mode and useJvmChmod` is just ignored.
Deprecate this setting so it can be removed from future versions.