Add ability to specify module version for a modular JAR file
Java 9 introduces modules and modular JAR files and it
allows you to specify version and main class for the
module. This metadata is stored inside the module descriptor -
the `module-info.class` file. Unfortunately the module
declaration(`module-info.java`) syntax does not allow those
values to be specified. This means that byte code manipulation
is required. Extend `JarArchiver` with methods to assist
with this task when modular JAR files are created.

The basic idea is to use `ModuleConfiguration` to set values
for the version and main class (in the future it may contain
other configuration such as if the module should be verified).
Instance of the class should be passed to
`JarArchiver#setModuleConfiguration` and when the JAR file
is created its module descriptor will be extended with the
specified values.

The byte code manipulation is done by the
`ModuleDescriptorExtender` class, but it is not part of the public
API and should not be used by clients of the Plexus Archiver.
Also it is worth mentioning that a new dependency is introduced
to help with the actual byte code manipulation - ASM.

This commit introduces only the ability to specify module version.
Subsequent commit will introduce the ability to specify
the main class as well.