Fix the way unit tests modify the timestamp of a file
The current approach is to copy the file until it gets a timestamp
newer than a given reference value.

Java 7 provides a better method to modify a file timestamp -
`Files.setLastModifiedTime`. Use it instead the current approach.
Not only it is better suited for the task, but Plexus Utils 3.1.0
no longer updates the timestamp of copied files and the current
approach is no longer working one - the tests are running
in endless loop.