Fix ZIP entries last modification time is rounded down
Zip archives store file modification times with a granularity of
two seconds, so the times will either be rounded up or down.
Plexus Archiver chooses to round up, but it assumes that Java 7
round the seconds down and Java 8 round them up.
In reality that is not true. On Java 8 java.util.zip.ZipEntry.getTime()
returns the same value set with java.util.zip.ZipEntry.setTime(), but
that does not matter as the last modification time is again rounded down
(if the seconds are odd value) when the entry is added to the archive.
As a result Plexus Archiver rounds the seconds up or Java 7 and
rounds them down on Java 8 or newer.

Remove the Java version check so Plexus Archiver rounds up the
last modification time on both Java 7 and Java 8+.

Fixes #95