Modify ModularJarArchiver to use the manifest main class as default
If the main class for a module is not explicitly set, lets use
the one specified in the manifest file. That would simplify the
creation of modular JAR files - clients should just set
the manifest main class as in non-modular JAR file and
everything will work out of the box. Also it will allow
`ModularJarArchiver` to be used as drop-in replacement of
`JarArchiver` without any additional configuration.