Tests for proper use of unicode path extra fields in zip unarchiving.
One of them currently failing due to a bug, to be fixed in the next commit.

Two test files included, one with the EFS bit (language encoding flag) set,
the other without. Only efsclear.zip is used in tests because the zip spec
does not specify which prevails when both the bit is set and an extra field
is present - plexus-archiver and all other unarchivers I've tried ignore
the extra field in that case.