Remove native `chmod` tool execution in tests
`Runtime.getRuntime().exec( "chmod 440 " + ` causes
security warnings because it concatenates arguments
(there is a risk of argument injection).

The argument does not come from untrusted source,
but as we can use Java to change a file mode, it is
better to remove the `chmod` invocation altogether.
`chmod` was necessary before Java 7 as there
was no support for changing file modes.